:recycle: Refactor Code - Exercício 03
